backend/requirements.txt
psycopg2-binary==2.9.9
gunicorn==21.2.0
Flask==3.0.3
Flask-Cors==5.0.0
